Move showScNewForm into separate file
