+++ src/applications/appeals/995/tests/utils/livingSituation.unit.spec.js
@@ -1,20 +1,11 @@
 import { expect } from 'chai';
 
 import {
-  showScNewForm,
   hasHousingRisk,
   hasOtherHousingRisk,
 } from '../../utils/livingSituation';
 
 import { SC_NEW_FORM_DATA } from '../../constants';
-
-describe('showScNewForm', () => {
-  it('should return expected value', () => {
-    expect(showScNewForm({ [SC_NEW_FORM_DATA]: undefined })).to.be.undefined;
-    expect(showScNewForm({ [SC_NEW_FORM_DATA]: true })).to.be.true;
-    expect(showScNewForm({ [SC_NEW_FORM_DATA]: false })).to.be.false;
-  });
-});
 
 describe('hasHousingRisk', () => {
   it('should return expected value', () => {
