+++ src/applications/appeals/995/utils/livingSituation.js
@@ -1,6 +1,4 @@
-import { SC_NEW_FORM_DATA } from '../constants';
-
-export const showScNewForm = formData => formData[SC_NEW_FORM_DATA];
+import { showScNewForm } from './toggle';
 
 export const hasHousingRisk = formData =>
   showScNewForm(formData) && formData.housingRisk;
