+++ src/applications/appeals/995/utils/toggle.js
@@ -1,0 +1,3 @@
+import { SC_NEW_FORM_DATA } from '../constants';
+
+export const showScNewForm = formData => formData[SC_NEW_FORM_DATA];
